# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CTM API Docs",
    "version": "1.0.1",
    "description": "CTM all endpoints"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/collection/nft": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "RequestBody",
            "in": "body",
            "required": true,
            "description": "Required request body to add nft to collection",
            "schema": {
              "type": "object",
              "properties": {
                "collectionId": {
                  "type": "string",
                  "example": "number"
                },
                "nftId": {
                  "type": "string",
                  "example": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response body at code 200. If nft belongs to a label, the wallet address is not null.    Default value for isSellable and isRentable is false, when nft is put up for sale, these values ​​will be changed to true",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "collection": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "nft": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number",
                                "example": 1
                              },
                              "name": {
                                "type": "string",
                                "example": "Chase the moonlite"
                              },
                              "price": {
                                "type": "number",
                                "example": 1
                              },
                              "quantity": {
                                "type": "number",
                                "example": 1
                              },
                              "tokenInStock": {
                                "type": "number",
                                "example": 1
                              },
                              "tokenListed": {
                                "type": "number",
                                "example": 0
                              },
                              "description": {
                                "type": "string",
                                "example": "Chase the moonlite description"
                              },
                              "imageURL": {
                                "type": "string",
                                "example": "https://imageURL/0"
                              },
                              "musiURL": {
                                "type": "string",
                                "example": "https://musicURL/0"
                              },
                              "nftMintTime": {
                                "type": "string",
                                "example": "1970-01-01 00:00:00.000"
                              },
                              "nftListTime": {},
                              "contractId": {
                                "type": "number",
                                "example": 1
                              },
                              "contractAddress": {
                                "type": "string",
                                "example": "0x0000000000000000000000000000000000000001"
                              },
                              "creatorWalletAddress": {
                                "type": "string",
                                "example": "0x0000000000000000000000000000000000000002"
                              },
                              "tokenStandard": {
                                "type": "string",
                                "example": "ERC1155"
                              },
                              "labelWallet": {},
                              "isSellable": {
                                "type": "boolean",
                                "example": false
                              },
                              "isRentable": {
                                "type": "boolean",
                                "example": false
                              },
                              "rentPrice": {
                                "type": "number",
                                "example": 1
                              },
                              "numberOfRents": {
                                "type": "number",
                                "example": 0
                              },
                              "numberOfCurrentRents": {
                                "type": "number",
                                "example": 0
                              },
                              "userLikes": {
                                "type": "number",
                                "example": 0
                              },
                              "userWhoLiked": {
                                "type": "array",
                                "example": [],
                                "items": {}
                              },
                              "moderatorPoints": {
                                "type": "number",
                                "example": 0
                              },
                              "rankingPoints": {
                                "type": "number",
                                "example": 0
                              },
                              "highestRank": {
                                "type": "number",
                                "example": 0
                              },
                              "views": {
                                "type": "number",
                                "example": 0
                              },
                              "downloads": {
                                "type": "number",
                                "example": 0
                              },
                              "genre": {
                                "type": "number",
                                "example": 1
                              },
                              "isFlagged": {
                                "type": "boolean",
                                "example": false
                              },
                              "flagDescription": {
                                "type": "string",
                                "example": "Flag description"
                              },
                              "collections": {
                                "type": "array",
                                "example": [
                                  1
                                ],
                                "items": {
                                  "type": "number"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "collection": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number",
                          "example": 1
                        },
                        "nft": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "number",
                                "example": 1
                              },
                              "name": {
                                "type": "string",
                                "example": "Chase the moonlite"
                              },
                              "price": {
                                "type": "number",
                                "example": 1
                              },
                              "quantity": {
                                "type": "number",
                                "example": 1
                              },
                              "tokenInStock": {
                                "type": "number",
                                "example": 1
                              },
                              "tokenListed": {
                                "type": "number",
                                "example": 0
                              },
                              "description": {
                                "type": "string",
                                "example": "Chase the moonlite description"
                              },
                              "imageURL": {
                                "type": "string",
                                "example": "https://imageURL/0"
                              },
                              "musiURL": {
                                "type": "string",
                                "example": "https://musicURL/0"
                              },
                              "nftMintTime": {
                                "type": "string",
                                "example": "1970-01-01 00:00:00.000"
                              },
                              "nftListTime": {},
                              "contractId": {
                                "type": "number",
                                "example": 1
                              },
                              "contractAddress": {
                                "type": "string",
                                "example": "0x0000000000000000000000000000000000000001"
                              },
                              "creatorWalletAddress": {
                                "type": "string",
                                "example": "0x0000000000000000000000000000000000000002"
                              },
                              "tokenStandard": {
                                "type": "string",
                                "example": "ERC1155"
                              },
                              "labelWallet": {},
                              "isSellable": {
                                "type": "boolean",
                                "example": false
                              },
                              "isRentable": {
                                "type": "boolean",
                                "example": false
                              },
                              "rentPrice": {
                                "type": "number",
                                "example": 1
                              },
                              "numberOfRents": {
                                "type": "number",
                                "example": 0
                              },
                              "numberOfCurrentRents": {
                                "type": "number",
                                "example": 0
                              },
                              "userLikes": {
                                "type": "number",
                                "example": 0
                              },
                              "userWhoLiked": {
                                "type": "array",
                                "example": [],
                                "items": {}
                              },
                              "moderatorPoints": {
                                "type": "number",
                                "example": 0
                              },
                              "rankingPoints": {
                                "type": "number",
                                "example": 0
                              },
                              "highestRank": {
                                "type": "number",
                                "example": 0
                              },
                              "views": {
                                "type": "number",
                                "example": 0
                              },
                              "downloads": {
                                "type": "number",
                                "example": 0
                              },
                              "genre": {
                                "type": "number",
                                "example": 1
                              },
                              "isFlagged": {
                                "type": "boolean",
                                "example": false
                              },
                              "flagDescription": {
                                "type": "string",
                                "example": "Flag description"
                              },
                              "collections": {
                                "type": "array",
                                "example": [
                                  1
                                ],
                                "items": {
                                  "type": "number"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - User is not logged in"
          },
          "403": {
            "description": "Forbidden - wrong role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "collectionId": {
                    "example": "any"
                  },
                  "nftId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User",
          "Label"
        ],
        "description": "Deleting nft from collection, user must be creator",
        "parameters": [
          {
            "name": "RequestBody",
            "in": "body",
            "required": true,
            "description": "Required request body to add nft to collection",
            "schema": {
              "type": "object",
              "properties": {
                "collectionId": {
                  "type": "string",
                  "example": "number"
                },
                "nftId": {
                  "type": "string",
                  "example": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response body at code 200.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "successfully added nft to collection (NFTs: [1, 2])"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "successfully added nft to collection (NFTs: [1, 2])"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - User is not logged in"
          },
          "403": {
            "description": "Forbidden - wrong role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "collectionId": {
                    "example": "any"
                  },
                  "nftId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}